Use a common request class for update and store where they are exactly the same
