+++ app/Http/Requests/CourseRequest.php
@@ -4,13 +4,15 @@ namespace App\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class MAUpdateRequest extends FormRequest {
+class CourseRequest extends FormRequest
+{
     /**
      * Determine if the user is authorized to make this request.
      *
      * @return bool
      */
-    public function authorize() {
+    public function authorize()
+    {
         return true;
     }
 
@@ -19,10 +21,11 @@ class MAUpdateRequest extends FormRequest {
      *
      * @return array
      */
-    public function rules() {
+    public function rules()
+    {
         return [
-            'anforderung' => 'required',
-            'killer' => 'boolean',
+            'name' => 'required',
+            'kursnummer' => '',
         ];
     }
 }
