+++ app/Http/Requests/QKRequest.php
@@ -4,13 +4,15 @@ namespace App\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class MAStoreRequest extends FormRequest {
+class QKRequest extends FormRequest
+{
     /**
      * Determine if the user is authorized to make this request.
      *
      * @return bool
      */
-    public function authorize() {
+    public function authorize()
+    {
         return true;
     }
 
@@ -19,10 +21,10 @@ class MAStoreRequest extends FormRequest {
      *
      * @return array
      */
-    public function rules() {
+    public function rules()
+    {
         return [
-            'anforderung' => 'required',
-            'killer' => 'boolean',
+            'quali_kategorie' => 'required',
         ];
     }
 }
